Add files via upload
Humans can also read and understand texts that contain spelling errors. This works (with common and well-known words) to the extent that letters that are almost completely mixed up can still be read:
You can not read your Txet, oedr? *

Implement a program that scrambles a given text in such a way that the first and last letters of the words are still in the correct position, but the letters in between are output in a sorted manner.